% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile_build.R
\name{build}
\alias{build}
\title{Build the supplied dockerfile object into a dockerimage}
\usage{
build(x, image, builddir = ".", verbose = TRUE)
}
\arguments{
\item{x}{a dockerfile object or path to a dockerfile}

\item{image}{name of the output image username/image[:tag]}

\item{builddir}{directory to build the docker image in. defaults to the current working directory}

\item{verbose}{should the command line results be printed to the R console?}
}
\value{
dockerfile
}
\description{
Build the dockerfile into a docker image that can be then deployed.
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1
dockerfile() \%>\%
  from("rocker/shiny:3.6.1")
}
\seealso{
Other dockerfile: \code{\link{cmd}}, \code{\link{copy}},
  \code{\link{expose}}, \code{\link{from}},
  \code{\link{install_r_lib}}, \code{\link{install}},
  \code{\link{run}}, \code{\link{update}},
  \code{\link{write_dockerfile}}
}
\concept{dockerfile}
