% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docker.R
\name{check_docker}
\alias{check_docker}
\title{Check if `docker` is in execution path}
\usage{
check_docker()
}
\value{
TRUE/FALSE
}
\description{
Check if the docker executable is available in the
   computers path. Not necessary for setting up dockerfile's, but expected to pass if
   you are expecting to take advantage of {dockyards} docker interactions.
}
\examples{
check_docker()
}
