% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_docker.R
\name{check_docker_imagename}
\alias{check_docker_imagename}
\title{Check image name format}
\usage{
check_docker_imagename(imagename)
}
\arguments{
\item{imagename}{character vector of name(s) of the output image. Should follow username/image[:tag] format}
}
\value{
TRUE/FALSE
}
\description{
Checks if the image name is a valid format
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1
check_docker_imagename("rocker/shiny:3.6.1")
check_docker_imagename("rocker/shiny")

}
