% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile_obj.R
\name{commands}
\alias{commands}
\title{Get the list of commands}
\usage{
commands(df)
}
\arguments{
\item{df}{a dockerfile object}
}
\value{
character vector
}
\description{
Get the list of commands from the dockerfile object.
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1
df <- dockerfile() \%>\%
  from("rocker/shiny:3.6.1") \%>\%
  run("echo 'hello world'")

commands(df)

}
