% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{copy}
\alias{copy}
\alias{copy_from_url}
\alias{add}
\title{Copy files into docker image}
\usage{
copy(df, source, dir)

copy_from_url(df, source, file)

add(df, source, dir)
}
\arguments{
\item{df}{a dockerfile object from `dockerfile()`}

\item{source}{file to be copied from your local computer or URL to use}

\item{dir}{location the file is to be copied to}

\item{file}{the output file to save the URL document as}
}
\description{
Command to copy files in the docker image during building the image.

Command to copy files from a URL into the docker image during building the image.

Command to ADd files from local computer into the docker image during building the image.
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1, update all existing software
# and install git and curl. Then initialize Rstudio server, and copy in a config file

dockerfile() \%>\%
 from("rocker/r-ver:devel") \%>\%
 update() \%>\%
 install("sudo","gdebi","pandoc","pandoc-citeproc",
         "libcurl4-gnutls-dev","libcairo2-dev",
         "libxtdev","wget") \%>\%
  run("wget --no-verbose
       https://download3.rstudio.org/ubuntu-14.04/x86_64/VERSION
       -O version.txt") \%>\%
  run("VERSION=$(cat version.txt)") \%>\%
  run("wget --no-verbose
       https://download3.rstudio.org/ubuntu-14.04/x86_64/shiny-server-$VERSION-amd64.deb
        -O ss-latest.deb") \%>\%
  run("gdebi -n ss-latest.deb") \%>\%
  run("rm -f version.txt ss-latest.deb") \%>\%
  run(". /etc/environment") \%>\%
  install_r_lib("shiny","rmarkdown") \%>\%
  run("cp -R /usr/local/lib/R/site-library/shiny/examples/* /srv/shiny-server/") \%>\%
  expose(3838) \%>\%
  copy("shiny-server.sh", "/usr/bin/shiny-server.sh")

# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1, update all existing software
# and install git and curl. Then initialize Rstudio server, and copy in a config file

dockerfile() \%>\%
 from("rocker/r-ver:devel") \%>\%
 update() \%>\%
 install("sudo","gdebi","pandoc","pandoc-citeproc",
         "libcurl4-gnutls-dev","libcairo2-dev",
         "libxtdev","wget","curl") \%>\%
  run("wget --no-verbose
      https://download3.rstudio.org/ubuntu-14.04/x86_64/VERSION
       -O version.txt") \%>\%
  run("VERSION=$(cat version.txt)") \%>\%
  run("wget --no-verbose
       https://download3.rstudio.org/ubuntu-14.04/x86_64/shiny-server-$VERSION-amd64.deb
       -O ss-latest.deb") \%>\%
  run("gdebi -n ss-latest.deb") \%>\%
  run("rm -f version.txt ss-latest.deb") \%>\%
  run(". /etc/environment") \%>\%
  install_r_lib("shiny","rmarkdown") \%>\%
  run("cp -R /usr/local/lib/R/site-library/shiny/examples/* /srv/shiny-server/") \%>\%
  expose(3838) \%>\%
  copy_from_url("https://raw.githubusercontent.com/rocker-org/shiny/master/shiny-server.sh",
                "/usr/bin/shiny-server.sh")

# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1, update all existing software
# and install git and curl. Then initialize Rstudio server, and copy in a config file

dockerfile() \%>\%
 from("rocker/r-ver:devel") \%>\%
 update() \%>\%
 install("sudo","gdebi","pandoc","pandoc-citeproc",
         "libcurl4-gnutls-dev","libcairo2-dev",
         "libxtdev","wget","curl") \%>\%
  run("wget --no-verbose
       https://download3.rstudio.org/ubuntu-14.04/x86_64/VERSION
       -O version.txt") \%>\%
  run("VERSION=$(cat version.txt)") \%>\%
  run("wget --no-verbose
       https://download3.rstudio.org/ubuntu-14.04/x86_64/shiny-server-$VERSION-amd64.deb
       -O ss-latest.deb") \%>\%
  run("gdebi -n ss-latest.deb") \%>\%
  run("rm -f version.txt ss-latest.deb") \%>\%
  run(". /etc/environment") \%>\%
  install_r_lib("shiny","rmarkdown") \%>\%
  run("cp -R /usr/local/lib/R/site-library/shiny/examples/* /srv/shiny-server/") \%>\%
  expose(3838) \%>\%
  copy_from_url("https://raw.githubusercontent.com/rocker-org/shiny/master/shiny-server.sh",
                "/usr/bin/shiny-server.sh") \%>\%
  add("/my/local/shiny/folder/","/srv/shiny-server/")

}
\seealso{
Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{expose}}, \code{\link{from}},
  \code{\link{install_r_lib_version}},
  \code{\link{install_r_lib}}, \code{\link{install}},
  \code{\link{run}}, \code{\link{update}},
  \code{\link{write_dockerfile}}

Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{expose}}, \code{\link{from}},
  \code{\link{install_r_lib_version}},
  \code{\link{install_r_lib}}, \code{\link{install}},
  \code{\link{run}}, \code{\link{update}},
  \code{\link{write_dockerfile}}

Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{expose}}, \code{\link{from}},
  \code{\link{install_r_lib_version}},
  \code{\link{install_r_lib}}, \code{\link{install}},
  \code{\link{run}}, \code{\link{update}},
  \code{\link{write_dockerfile}}
}
\concept{dockerfile}
