% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_conn.R
\name{docker_connection}
\alias{docker_connection}
\title{Define the connection to the docker container}
\usage{
docker_connection(name)
}
\arguments{
\item{name}{the name of the docker image}
}
\value{
dockefile object
}
\description{
Define the connection to the docker container
}
\details{
This object defines the connection to the docker image when it can accept
external inputs
}
