% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker_kill}
\alias{docker_kill}
\title{kill a docker container}
\usage{
docker_kill(x)
}
\arguments{
\item{x}{either the name of the docker container or a docker_connection}
}
\description{
kill a currently running docker container
}
