% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker_pause}
\alias{docker_pause}
\title{Pause a docker container to maintain its state}
\usage{
docker_pause(x)
}
\arguments{
\item{x}{either the name of the docker container or a docker_connection}
}
\description{
Pause a docker container from the list of docker containers
}
