% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker_rm}
\alias{docker_rm}
\title{Remove a docker container from the listed containers}
\usage{
docker_rm(x)
}
\arguments{
\item{x}{either the name of the docker container or a docker_connection}
}
\description{
Remove a docker container from the list of docker containers
}
