% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker_run}
\alias{docker_run}
\title{Initialize a docker container from an image}
\usage{
docker_run(image, name, ports, mountpoints, docker_run_args)
}
\arguments{
\item{image}{Repository address in the format username/image[:tag]}

\item{name}{name of the docker container. can be left empty}

\item{ports}{what port to expose and connect to in the docker container
format is port:port}

\item{mountpoints}{what folder to expose to the docker container, where to
connect to it within the docker container. format is dir:dir}

\item{docker_run_args}{any additional arguments to be passed to the `docker run` command.}
}
\value{
docker_conn object
}
\description{
Initialize a docker image into a docker container. Generates a
    dockercon object to help monitor status of docker container.
}
\examples{
\dontrun{
# Start a docker container from the rocker/r-ver:devel image
d_conn<-docker_run(image = "rocker/r-ver:devel", name = "testimage")
}
}
