% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker_start}
\alias{docker_start}
\title{Start a docker container to maintain its state}
\usage{
docker_start(x)
}
\arguments{
\item{x}{either the name of the docker container or a docker_connection}
}
\description{
Start a stopped docker container
}
