% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker_stop}
\alias{docker_stop}
\title{Stop a docker container from running}
\usage{
docker_stop(x)
}
\arguments{
\item{x}{either the name of the docker container or a docker_connection}
}
\description{
Stop a currently running docker container
}
