% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_run.R
\name{docker_unpause}
\alias{docker_unpause}
\title{unpause a docker container}
\usage{
docker_unpause(x)
}
\arguments{
\item{x}{either the name of the docker container or a docker_connection}
}
\description{
unpause a paused docker container
}
