% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile_obj.R
\name{dockerfile}
\alias{dockerfile}
\title{Specify dockerfile to base build off of}
\usage{
dockerfile(path = tempfile())
}
\arguments{
\item{path}{path to either existing dockerfile to base off of, or where the output will be}
}
\value{
dockefile object
}
\description{
This function should be used with the dockercommands family of functions to generate
a dockerfile to be used for building a new docker image
}
\examples{
\dontrun{
df <- dockerfile()
df2 <- dockerfile("path/to/existing/dockerfile")
}
}
