% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{install}
\alias{install}
\title{Apt-get install software}
\usage{
install(df, ...)
}
\arguments{
\item{df}{a dockerfile object from `dockerfile()`}

\item{...}{The software to install to the docker image}
}
\value{
dockerfile
}
\description{
Add software to install to docker image via `apt-get`
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1, update all existing software
# and install git and curl.

dockerfile() \%>\%
  from("rocker/r-ver:devel") \%>\%
  update() \%>\%
  install("sudo","gdebi","pandoc","pandoc-citeproc",
          "libcurl4-gnutls-dev","libcairo2-dev",
          "libxtdev","wget")
}
\seealso{
Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{copy}}, \code{\link{expose}},
  \code{\link{from}}, \code{\link{install_r_lib}},
  \code{\link{run}}, \code{\link{save}},
  \code{\link{update}}
}
\concept{dockerfile}
