% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{install_r_lib}
\alias{install_r_lib}
\title{Install packages for R}
\usage{
install_r_lib(df, ...)
}
\arguments{
\item{df}{a dockerfile object from `dockerfile()`}

\item{...}{The R packages to install to the docker image}
}
\description{
install packages from the default CRAN location for R
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1, update all existing software
# and install git and curl.

dockerfile() \%>\%
 from("rocker/r-ver:devel") \%>\%
 update() \%>\%
 install("sudo","gdebi","pandoc","pandoc-citeproc",
         "libcurl4-gnutls-dev","libcairo2-dev",
         "libxtdev","wget") \%>\%
  install_r_lib("tidyverse")
}
\seealso{
Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{copy}}, \code{\link{expose}},
  \code{\link{from}}, \code{\link{install}},
  \code{\link{run}}, \code{\link{update}},
  \code{\link{write_dockerfile}}
}
\concept{dockerfile}
