% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{install_r_lib_version}
\alias{install_r_lib_version}
\title{Install specific package version for R}
\usage{
install_r_lib_version(df, package, version, new = FALSE)
}
\arguments{
\item{df}{a dockerfile object from `dockerfile()`}

\item{package}{The R package to install to the docker image as a character.}

\item{version}{The version R package to install to the docker image as a character.}

\item{new}{boolean indicating if the command should be added to the prior `Run`` line.
Useful for when breaking out lines in the docker image to skip time consuming library build steps.
defaults to FALSE.}
}
\description{
install packages from the default CRAN location for R
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1, update all existing software
# and install git and curl.

dockerfile() \%>\%
 from("rocker/r-ver:devel") \%>\%
 update() \%>\%
 install("sudo","gdebi","pandoc","pandoc-citeproc",
         "libcurl4-gnutls-dev","libcairo2-dev",
         "libxtdev","wget") \%>\%
  install_r_lib("tidyverse") \%>\%
  install_r_lib_version("ggplot2", version = "0.9.1")
}
\seealso{
Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{copy}}, \code{\link{expose}},
  \code{\link{from}}, \code{\link{install_r_lib}},
  \code{\link{install}}, \code{\link{run}},
  \code{\link{update}}, \code{\link{write_dockerfile}}
}
\concept{dockerfile}
