% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_conn.R
\name{list_containers}
\alias{list_containers}
\title{List all docker containers available to docker daemon}
\usage{
list_containers()
}
\value{
data.frame of docker containers
}
\description{
List all docker containers available to docker daemon
}
\details{
This function asks the docker daemon to list all the available
   docker comtainers that are available.
}
