% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docker_conn.R
\name{list_images}
\alias{list_images}
\title{List all locally exising docker images available to docker daemon}
\usage{
list_images()
}
\value{
data.frame of docker images
}
\description{
List all locally exising docker images available to docker daemon
}
\details{
This function asks the docker daemon to list all the available
   docker images that are available.
}
