% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile_build.R
\name{save}
\alias{save}
\title{RUN command in docker image}
\usage{
save(df, path, overwrite = FALSE)
}
\arguments{
\item{df}{a dockerfile object from `dockerfile()`}

\item{path}{path to where the dockerfile will be saved at}

\item{overwrite}{should the current dockefile be overwritten?}
}
\description{
Command to cope files  in the docker image during building the image.
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1, update all existing software
# and install git and curl. Then initialize Rstudio server, and copy in a config file

new_dockerfile <- tempfile(pattern = "new_dockerfile_", fileext = "")


df <- dockerfile() \%>\%
  from("rocker/shiny:3.6.1") \%>\%
  update() \%>\%
  install("git", "libcurl4-openssl-dev") \%>\%
  install_r_lib("rcurl", "dplyr") \%>\%
  run("wget --no-check-certificate
   https://raw.githubusercontent.com/rocker-org/rstudio-daily/master/latest.R") \%>\%
  run("Rscript latest.R") \%>\%
  run("dpkg -i rstudio-server-daily-amd64.deb") \%>\%
  copy("userconf.sh", "/etc/cont-init.d/userconf") \%>\%
  save(new_dockerfile)
}
\seealso{
Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{copy}}, \code{\link{expose}},
  \code{\link{from}}, \code{\link{install_r_lib}},
  \code{\link{install}}, \code{\link{run}},
  \code{\link{update}}
}
\concept{dockerfile}
