% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{update}
\alias{update}
\title{Run update docker image}
\usage{
update(df)
}
\arguments{
\item{df}{a dockerfile object from `dockerfile()`}
}
\value{
dockerfile
}
\description{
Add command when creating docker image to update installed software
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1
dockerfile() \%>\%
  from("rocker/r-ver:devel") \%>\%
  update()
}
\seealso{
Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{copy}}, \code{\link{expose}},
  \code{\link{from}}, \code{\link{install_r_lib}},
  \code{\link{install}}, \code{\link{run}},
  \code{\link{write_dockerfile}}
}
\concept{dockerfile}
