% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{from}
\alias{from}
\title{Describe the base docker image to use}
\usage{
from(df, image = "rocker/r-base")
}
\arguments{
\item{df}{a dockerfile object from `dockerfile()`}

\item{image}{Repository address in the format username/image[:tag]}
}
\value{
dockerfile
}
\description{
Creating a docker image in a dockerfile requires specification
   of the base image to be creating your new docker image off of. {dockyard}
   assumes the user is basing their image off the latest available image
   from rocker/r-base, but it can be changed at any time.
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1
dockerfile() \%>\%
  from("rocker/r-ver:devel")
}
\seealso{
Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{copy}}, \code{\link{expose}},
  \code{\link{install_r_lib}}, \code{\link{install}},
  \code{\link{run}}, \code{\link{update}},
  \code{\link{write_dockerfile}}
}
\concept{dockerfile}
