% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dockerfile.R
\name{run}
\alias{run}
\title{RUN command in docker image}
\usage{
run(df, cmd, new = FALSE)
}
\arguments{
\item{df}{a dockerfile object from `dockerfile()`}

\item{cmd}{command to be run in the docker image during build}

\item{new}{boolean indicating if the command should be added to the prior line.
Useful for when breaking out lines in the docker image to skip time consuming build steps.
 defaults to FALSE.}
}
\description{
Command to be run in the docker image during building the image.
}
\examples{
# Start a dockerfile based off of the rocker/shiny image to generate a
# shiny server using R version 3.6.1, update all existing software
# and install git and curl. Then initialize Rstudio server

dockerfile() \%>\%
 from("rocker/r-ver:devel") \%>\%
 update() \%>\%
 install("sudo","gdebi","pandoc","pandoc-citeproc",
         "libcurl4-gnutls-dev","libcairo2-dev",
         "libxtdev","wget") \%>\%
  run("wget --no-verbose
       https://download3.rstudio.org/ubuntu-14.04/x86_64/VERSION
       -O version.txt") \%>\%
  run("VERSION=$(cat version.txt)") \%>\%
  run("wget --no-verbose
       https://download3.rstudio.org/ubuntu-14.04/x86_64/shiny-server-$VERSION-amd64.deb
       -O ss-latest.deb") \%>\%
  run("gdebi -n ss-latest.deb") \%>\%
  run("rm -f version.txt ss-latest.deb") \%>\%
  run(". /etc/environment") \%>\%
  install_r_lib("shiny","rmarkdown") \%>\%
  run("cp -R /usr/local/lib/R/site-library/shiny/examples/* /srv/shiny-server/")

}
\seealso{
Other dockerfile: \code{\link{build}}, \code{\link{cmd}},
  \code{\link{copy}}, \code{\link{expose}},
  \code{\link{from}}, \code{\link{install_r_lib_version}},
  \code{\link{install_r_lib}}, \code{\link{install}},
  \code{\link{update}}, \code{\link{write_dockerfile}}
}
\concept{dockerfile}
